requirements.txt
aiogram>=3.4.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
apscheduler>=3.10.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
